requirements.txt
sqlmodel
google-generativeai
pandas
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
